tc4 - pre Div operator fix
