+++ test/runtests.jl
@@ -1,3 +1,4 @@
 using Dates: include
 include("cplstate_interface.jl")
 include("cplsolver.jl")
+#include("experiments_ClimaCore.jl")
